requirements.txt
boto3~=1.28.0
click~=8.1.0
pydantic~=2.1.0
pyhumps~=3.8.0
python-dotenv~=1.0.0
spotipy~=2.23.0
sqlalchemy~=2.0.0
